requirements.txt
asgiref==3.8.1
black==24.4.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
firebase-admin==6.5.0
gcloud==0.18.3
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
jwcrypto==1.5.6
jwt==1.3.1
msgpack==1.0.8
mypy-extensions==1.0.0
oauth2client==4.1.3
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.9.0
pyparsing==3.1.2
Pyrebase4==4.8.0
pyrebase5==5.0.1
python-dotenv==1.0.1
python-jwt==4.1.0
requests==2.31.0
requests-toolbelt==0.10.1
rsa==4.9
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.18
